Add executable and command-line options
Create an executable for the gem using the trollop library
for command-line parsing.

Add basic command-line options for database connection and CSV
output. The original script has some other options but I am not
sure yet if these will be used. If it turns out that they are
needed then we can add them later.

Also updated the README with usage information based on the
above as well as a little information about the origins of this
project.