Add load path to executable
Add the gem's library directory to the ruby load-path. This
allows us to run the command without having to isntall the
gem which is useful in development.

Also move the trollop require above the bsf version require
just because it makes more sense to require the general stuff
first. At least that is my opinion