Move argument parsing into Command class.
The Trollop command-line parsing library allows us to specify
a variable containing the command-line arguments. Therefore to
keep the executable code as simple as possible and aid
testability we will move the command-line parsing into the
Bsf::Scraper::Command class.

The trollop code was initially in the executable file because
I did not know about this possibility and thought it was
hard-coded to parse ARGV