Create Fund table
Create a table called "funds" which will store all the scraped
data. Include created_at and updated_at timestamps.

Currently a private method of the Command class. We may want
to extract this (along with the database connection method) into
a separate class.

No special tests for this since it is basically 100% Sequel
library code and we trust that library's tests.

The size values of the first 4 fields are probably too big.
Resize them later when we have a better idea of the size of
the strings that will sit in them.