Add FundIndexer and Fund classes
Add the FundIndexer class whos job it is to visit the index
pages, parse them and create the funds in the database.

Add the Fund class which is a Model class linking to the
database.

To keep things fast and avoid hitting the Bloomberg site to much
We will use the VCR / Webmock libraries to stub out requests made
and replace the PostgreSQL database with an In-memory SQLite3
database when running tests