Add spidering code to FundIndexer
Add code which will follow the 'next' link on the fund pages
to get the next page for crawling.

Full tests have not been set-up in this commit due to time
required to create good sample test data in the form of modified
pages. The pages to be modified (and then served up by webmock)
are in the spec/fixtures/webmock_responses/fund_index_pages
directory.

We will remove the VCR code for the moment since that returns
and caches the actual responses from the bloomberg site which
means we would be crawling and parse 154 pages for each test.
Even with cached responses this takes too long to a unit test
plus the cached responses file is 4.8MB

Not sure if VCR will be used in the future for performance
testing so lets leave it in for the moment, we can always remove
it later.