Add option to skip indexing of funds
Add the option to skip indexing of funds. This index probably
changes very infrequently so we will give the script an option
to skip the indexing step.

This will allows us to setup a daily cronjob without the
indexing step and a weekly cronjob with the indexing step.

This will allow the script to execute more quickly.