Add yahoo finance profile page parser
Add the basic Yahoo Finance fund profile page parser with some
(not all) data parsing methods added.

Have tested using a basic Yahoo Finance page, testing all the
variations of those pages will be a bit of a faff given current
design.

While aiming for 100% coverage is nice in this case the
time-sink does not merit the extra work at the moment. Later,
when the script is finished and working, we can refactor this
piece of code in such a way that testing will be easier.