Complete Yahoo Finance fund profile page parser
This class scrapes all the fields that the original script does
from the profile page.

Due to the rather hacky method of getting the data by using
nokogiri to isolate part of the page we are interested in then
regexp'ing the rest the code parses a page in about 30ms on a
Intel Core i7-2600 CPU @ 3.40GHz with MRI 1.9.3

Because the original script is a bit of a PITA to setup I
cannot compare this to the original.

While working on this code I also took the opportunity to
decrypt some of the database field names. I have added them
as comments in database creation method although ideally we
would refactor the real database field names to be more
descriptive.